Move useInvoice and useQrPayment into own files
