+++ components/use-invoice.js
@@ -1,12 +1,10 @@
-import { useCallback } from 'react'
 import { useApolloClient, useMutation } from '@apollo/client'
-import { CANCEL_INVOICE, INVOICE } from '@/fragments/wallet'
-import Invoice from '@/components/invoice'
-import { useShowModal } from './modal'
-import { InvoiceCanceledError, InvoiceExpiredError } from '@/wallets/errors'
+import { useCallback } from 'react'
 import { RETRY_PAID_ACTION } from '@/fragments/paidAction'
+import { INVOICE, CANCEL_INVOICE } from '@/fragments/wallet'
+import { InvoiceExpiredError, InvoiceCanceledError } from '@/wallets/errors'
 
-export const useInvoice = () => {
+export default function useInvoice () {
   const client = useApolloClient()
   const [retryPaidAction] = useMutation(RETRY_PAID_ACTION)
 
@@ -59,46 +57,4 @@ export const useInvoice = () => {
   }, [retryPaidAction])
 
   return { cancel, retry, isInvoice }
-}
-
-export const useQrPayment = () => {
-  const invoice = useInvoice()
-  const showModal = useShowModal()
-
-  const waitForQrPayment = useCallback(async (inv, walletError,
-    {
-      keepOpen = true,
-      cancelOnClose = true,
-      persistOnNavigate = false,
-      waitFor = inv => inv?.satsReceived > 0
-    } = {}
-  ) => {
-    return await new Promise((resolve, reject) => {
-      let paid
-      const cancelAndReject = async (onClose) => {
-        if (!paid && cancelOnClose) {
-          const updatedInv = await invoice.cancel(inv).catch(console.error)
-          reject(new InvoiceCanceledError(updatedInv))
-        }
-        resolve(inv)
-      }
-      showModal(onClose =>
-        <Invoice
-          id={inv.id}
-          modal
-          description
-          status='loading'
-          successVerb='received'
-          walletError={walletError}
-          waitFor={waitFor}
-          onExpired={inv => reject(new InvoiceExpiredError(inv))}
-          onCanceled={inv => { onClose(); reject(new InvoiceCanceledError(inv, inv?.actionError)) }}
-          onPayment={(inv) => { paid = true; onClose(); resolve(inv) }}
-          poll
-        />,
-      { keepOpen, persistOnNavigate, onClose: cancelAndReject })
-    })
-  }, [invoice])
-
-  return waitForQrPayment
 }
