+++ components/use-paid-mutation.js
@@ -1,6 +1,7 @@
 import { useApolloClient, useLazyQuery, useMutation } from '@apollo/client'
 import { useCallback, useState } from 'react'
-import { useInvoice, useQrPayment } from './payment'
+import useQrPayment from '@/components/use-qr-payment'
+import useInvoice from '@/components/use-invoice'
 import { InvoiceCanceledError, InvoiceExpiredError, WalletError, WalletPaymentError } from '@/wallets/errors'
 import { GET_PAID_ACTION } from '@/fragments/paidAction'
 import { useWalletPayment } from '@/wallets/payment'
